refactor: member 관련 DTO 구획화
