+++ src/main/java/com/nhnacademy/member_server/dto/request/member/MemberCreateRequest.java
@@ -1,14 +1,19 @@
-package com.nhnacademy.member_server.dto.request;
+package com.nhnacademy.member_server.dto.request.member;
 
+import com.nhnacademy.member_server.entity.Gender;
+import com.nhnacademy.member_server.entity.Role;
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotNull;
 import java.time.LocalDate;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
+import lombok.RequiredArgsConstructor;
 
 @Getter
 @AllArgsConstructor
-public class SignupRequest {
+@RequiredArgsConstructor
+public class MemberCreateRequest {
     @NotBlank
     String loginId;
 
@@ -24,5 +29,11 @@ public class SignupRequest {
     @Email
     private String email;
 
+    @NotNull(message = "성별은 필수입니다.")
+    private Gender gender;
+
     private LocalDate birthDate;
+
+
+    private Role role;
 }
