+++ src/main/java/com/nhnacademy/member_server/dto/request/member/SignupRequest.java
@@ -1,19 +1,14 @@
-package com.nhnacademy.member_server.dto.request;
+package com.nhnacademy.member_server.dto.request.member;
 
-import com.nhnacademy.member_server.entity.Gender;
-import com.nhnacademy.member_server.entity.Role;
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.NotNull;
 import java.time.LocalDate;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
-import lombok.RequiredArgsConstructor;
 
 @Getter
 @AllArgsConstructor
-@RequiredArgsConstructor
-public class MemberCreateRequest {
+public class SignupRequest {
     @NotBlank
     String loginId;
 
@@ -29,11 +24,5 @@ public class MemberCreateRequest {
     @Email
     private String email;
 
-    @NotNull(message = "성별은 필수입니다.")
-    private Gender gender;
-
     private LocalDate birthDate;
-
-
-    private Role role;
 }
