+++ src/main/java/com/nhnacademy/member_server/dto/response/member/AddressListResponse.java
@@ -1,4 +1,4 @@
-package com.nhnacademy.member_server.dto.response;
+package com.nhnacademy.member_server.dto.response.member;
 
 import java.util.List;
 import lombok.AllArgsConstructor;
