+++ src/main/java/com/nhnacademy/member_server/service/AddressService.java
@@ -1,8 +1,8 @@
 package com.nhnacademy.member_server.service;
 
-import com.nhnacademy.member_server.dto.request.AddressRequest;
-import com.nhnacademy.member_server.dto.response.AddressListResponse;
-import com.nhnacademy.member_server.dto.response.AddressResponse;
+import com.nhnacademy.member_server.dto.request.member.AddressRequest;
+import com.nhnacademy.member_server.dto.response.member.AddressListResponse;
+import com.nhnacademy.member_server.dto.response.member.AddressResponse;
 
 public interface AddressService {
     public AddressResponse findDefaultAddress(Long memberId);
