+++ src/main/java/com/nhnacademy/member_server/service/impl/MemberServiceImpl.java
@@ -1,8 +1,8 @@
 package com.nhnacademy.member_server.service.impl;
 
-import com.nhnacademy.member_server.dto.request.MemberUpdateRequest;
-import com.nhnacademy.member_server.dto.response.MemberResponse;
-import com.nhnacademy.member_server.dto.response.SimpleMemberResponse;
+import com.nhnacademy.member_server.dto.request.member.MemberUpdateRequest;
+import com.nhnacademy.member_server.dto.response.member.MemberResponse;
+import com.nhnacademy.member_server.dto.response.member.SimpleMemberResponse;
 import com.nhnacademy.member_server.entity.Member;
 import com.nhnacademy.member_server.entity.Role;
 import com.nhnacademy.member_server.entity.Status;
