+++ src/test/java/com/nhnacademy/member_server/controller/MemberControllerTest.java
@@ -1,8 +1,8 @@
 package com.nhnacademy.member_server.controller;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.nhnacademy.member_server.dto.request.MemberUpdateRequest;
-import com.nhnacademy.member_server.dto.response.MemberResponse;
+import com.nhnacademy.member_server.dto.request.member.MemberUpdateRequest;
+import com.nhnacademy.member_server.dto.response.member.MemberResponse;
 import com.nhnacademy.member_server.entity.Gender;
 import com.nhnacademy.member_server.entity.MemberPrincipal;
 import com.nhnacademy.member_server.service.AuthService;
